[docs] Move constant files to constants folder for preval scoping
