+++ docs/components/DocumentationHeader.js
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import * as React from 'react';
 import { css } from 'react-emotion';
 
-import * as Constants from '~/common/constants';
+import * as Constants from '~/constants/theme';
 import AlgoliaSearch from '~/components/plugins/AlgoliaSearch';
 
 const STYLES_LOGO = css`
