+++ docs/components/base/details.js
@@ -1,7 +1,7 @@
 import styled, { keyframes, css } from 'react-emotion';
 
 import * as React from 'react';
-import * as Constants from '~/common/constants';
+import * as Constants from '~/constants/theme';
 
 const STYLES_DETAILS = css`
   margin-bottom: 0px;
