+++ docs/components/icons/CheckCircle.js
@@ -1,4 +1,4 @@
-import * as Constants from '~/common/constants';
+import * as Constants from '~/constants/theme';
 
 export const CheckCircle = ({ size = 24 }) => (
   <svg
