+++ docs/components/plugins/TableOfContentSection.js
@@ -1,7 +1,7 @@
 import styled, { keyframes, css } from 'react-emotion';
 import * as React from 'react';
 import BulletIcon from '~/components/icons/Bullet';
-import * as Constants from '~/common/constants';
+import * as Constants from '~/constants/theme';
 import { InlineCode } from '~/components/base/code';
 import { H4 } from '~/components/base/headings';
