+++ docs/global-styles/reset.js
@@ -1,4 +1,4 @@
-import * as Constants from '~/common/constants';
+import * as Constants from '~/constants/theme';
 
 export const globalReset = `
     html, body, div, span, applet, object, iframe,
